requirements.txt
pyinstaller
pdf2image~=1.17.0
pyzbar~=0.1.9
urllib3
kraken~=2.0.8
pillow~=10.2.0
numpy~=1.26.3